+++ CHANGELOG.md
@@ -8,7 +8,8 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 ## UNRELEASED
 
 - [#1139](https://github.com/ontola/atomic-server/issues/1139) AtomicServer can now create data without being dependent on a server! AtomicServer is now Local-First, using the new `did:ad` schema.
-- #584 Replace ureq with reqwest (async HTTP calls) 
+- #584 Replace ureq with reqwest (async HTTP calls)
+- #481 Drive scoped queries 
 - #420 Fix OTLP / OpenTelemetry, update docs from Jaeger to SigNoz, add metrics
 - [#590](https://github.com/ontola/atomic-server/issues/590) Get rid of the `SERVER_URL` env var, which makes moving & setup easier. All resources are now relative to the hosted domain, and AtomicServer can be available from multiple domains at once.
 - [#544](https://github.com/ontola/atomicdata-dev/atomic-server/issues/544) Stateless invites, using JWTs. Server setup now requires you to check the logs for the invite token.
@@ -24,8 +25,8 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - [#1024](https://github.com/ontola/atomic-server/issues/1024) Added URI datatype.
 - [#998](https://github.com/ontola/atomic-server/issues/998) Added YJS datatype.
 - [#851](https://github.com/ontola/atomic-server/issues/851) Deleting file resources now also deletes the file from the filesystem.
-BREAKING: [#1107](https://github.com/ontola/atomic-server/issues/1107) Named nested resources are no longer supported. Value::Resource and SubResource::Resource have been removed. If you need to include multiple resources in a response use an array.
-BREAKING: `store.get_resource_extended()` now returns a `ResourceResponse` instead of a `Resource` due to the removal of named nested resources. Use `.into()` or `.to_single()` to convert to a `Resource`.
+  BREAKING: [#1107](https://github.com/ontola/atomic-server/issues/1107) Named nested resources are no longer supported. Value::Resource and SubResource::Resource have been removed. If you need to include multiple resources in a response use an array.
+  BREAKING: `store.get_resource_extended()` now returns a `ResourceResponse` instead of a `Resource` due to the removal of named nested resources. Use `.into()` or `.to_single()` to convert to a `Resource`.
 
 ## [v0.40.2]